[WFLY-19966]: Remove hornetq-* jar dependencies from the base redistributable package.
 * Create a SPI for removing direct use of HornetQ code in the
   messaging-extension
 * Extract all HornetQ related code into legacy/hornetq/hornetq-extension
 * Create an HornetQ Feature Pack
 * Update the legacy integration test to use WildFly with HornetQ
   provisionned
 * Remove the HornetQ module from WildFly

Jira: https://issues.redhat.com/browse/WFLY-19966
