+++ legacy/hornetq/hornetq-extension/src/main/resources/META-INF/services/org.wildfly.extension.messaging.activemq.jms.legacy.LegacyConnectionFactory
@@ -1,0 +1,5 @@
+#
+# Copyright The WildFly Authors
+# SPDX-License-Identifier: Apache-2.0
+#
+org.wildfly.extension.messaging.activemq.jms.legacy.LegacyConnectionFactoryImpl
