+++ legacy/hornetq/hornetq-feature-pack/src/main/resources/modules/system/layers/base/org/hornetq/client/main/module.xml
