+++ legacy/hornetq/hornetq-feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/extension/hornetq/main/module.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+  ~ Copyright The WildFly Authors
+  ~ SPDX-License-Identifier: Apache-2.0
+  -->
+
+<module xmlns="urn:jboss:module:1.9" name="org.wildfly.extension.hornetq">
+
+    <properties>
+        <property name="jboss.api" value="private"/>
+    </properties>
+
+    <resources>
+        <artifact name="${org.wildfly:wildfly-hornetq-extension}"/>
+    </resources>
+
+
+     <provides>
+        <service name="org.wildfly.extension.messaging.activemq.jms.legacy.LegacyConnectionFactory">
+          <with-class name="org.wildfly.extension.messaging.activemq.jms.legacy.LegacyConnectionFactoryImpl"/>
+        </service>
+    </provides>
+
+    <dependencies>
+        <module name="java.management"/>
+        <module name="java.naming"/>
+        <module name="jakarta.transaction.api"/>
+        <module name="jakarta.jms.api"/>
+        <module name="jakarta.resource.api"/>
+        <module name="org.apache.activemq.artemis"/>
+        <module name="org.jboss.as.controller"/>
+        <module name="org.jboss.logging"/>
+        <module name="org.jboss.msc"/>
+        <module name="org.hornetq.client" export="true"/>
+        <module name="org.wildfly.extension.messaging-activemq"/>
+    </dependencies>
+</module>
