+++ messaging-activemq/subsystem/src/main/java/org/wildfly/extension/messaging/activemq/jms/legacy/LegacyConnectionFactoryService.java
@@ -2,29 +2,16 @@
  * Copyright The WildFly Authors
  * SPDX-License-Identifier: Apache-2.0
  */
-
 package org.wildfly.extension.messaging.activemq.jms.legacy;
+
 
 import static org.wildfly.extension.messaging.activemq.CommonAttributes.LEGACY;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import jakarta.jms.ConnectionFactory;
 
-import org.apache.activemq.artemis.api.config.ActiveMQDefaultConfiguration;
 import org.apache.activemq.artemis.core.server.ActiveMQServer;
-import org.hornetq.api.config.HornetQDefaultConfiguration;
-import org.hornetq.api.core.BroadcastEndpointFactoryConfiguration;
-import org.hornetq.api.core.DiscoveryGroupConfiguration;
-import org.hornetq.api.core.TransportConfiguration;
-import org.hornetq.api.core.UDPBroadcastGroupConfiguration;
-import org.hornetq.api.jms.HornetQJMSClient;
-import org.hornetq.api.jms.JMSFactoryType;
-import org.hornetq.core.remoting.impl.netty.TransportConstants;
-import org.hornetq.jms.client.HornetQConnectionFactory;
 import org.jboss.msc.service.Service;
 import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceController;
@@ -37,117 +24,22 @@ import org.jboss.msc.value.InjectedValue;
 import org.wildfly.extension.messaging.activemq.ActiveMQActivationService;
 import org.wildfly.extension.messaging.activemq.ActiveMQBroker;
 import org.wildfly.extension.messaging.activemq.jms.JMSServices;
-import org.wildfly.extension.messaging.activemq._private.MessagingLogger;
 
 /**
  * @author <a href="http://jmesnil.net/">Jeff Mesnil</a> (c) 2015 Red Hat inc.
  */
 public class LegacyConnectionFactoryService implements Service<ConnectionFactory> {
 
-    /**
-     * Map ActiveMQ parameters key (using CameCalse convention) to HornetQ parameter keys (using lisp-case convention)
-     */
-    private static final Map<String, String> PARAM_KEY_MAPPING = new HashMap<>();
-
-    static {
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.SSL_ENABLED_PROP_NAME,
-                TransportConstants.SSL_ENABLED_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.HTTP_ENABLED_PROP_NAME,
-                TransportConstants.HTTP_ENABLED_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.HTTP_CLIENT_IDLE_PROP_NAME,
-                TransportConstants.HTTP_CLIENT_IDLE_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.HTTP_CLIENT_IDLE_SCAN_PERIOD,
-                TransportConstants.HTTP_CLIENT_IDLE_SCAN_PERIOD);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.HTTP_REQUIRES_SESSION_ID,
-                TransportConstants.HTTP_REQUIRES_SESSION_ID);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.HTTP_UPGRADE_ENABLED_PROP_NAME,
-                TransportConstants.HTTP_UPGRADE_ENABLED_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.HTTP_UPGRADE_ENDPOINT_PROP_NAME,
-                TransportConstants.HTTP_UPGRADE_ENDPOINT_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.USE_SERVLET_PROP_NAME,
-                TransportConstants.USE_SERVLET_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.SERVLET_PATH,
-                TransportConstants.SERVLET_PATH);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.USE_NIO_PROP_NAME,
-                TransportConstants.USE_NIO_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.USE_NIO_GLOBAL_WORKER_POOL_PROP_NAME,
-                TransportConstants.USE_NIO_GLOBAL_WORKER_POOL_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.LOCAL_ADDRESS_PROP_NAME,
-                TransportConstants.LOCAL_ADDRESS_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.KEYSTORE_PROVIDER_PROP_NAME,
-                TransportConstants.KEYSTORE_PROVIDER_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.KEYSTORE_PATH_PROP_NAME,
-                TransportConstants.KEYSTORE_PATH_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.KEYSTORE_PASSWORD_PROP_NAME,
-                TransportConstants.KEYSTORE_PASSWORD_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.TRUSTSTORE_PROVIDER_PROP_NAME,
-                TransportConstants.TRUSTSTORE_PROVIDER_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.TRUSTSTORE_PATH_PROP_NAME,
-                TransportConstants.TRUSTSTORE_PATH_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.TRUSTSTORE_PASSWORD_PROP_NAME,
-                TransportConstants.TRUSTSTORE_PASSWORD_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.ENABLED_CIPHER_SUITES_PROP_NAME,
-                TransportConstants.ENABLED_CIPHER_SUITES_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.ENABLED_PROTOCOLS_PROP_NAME,
-                TransportConstants.ENABLED_PROTOCOLS_PROP_NAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.TCP_NODELAY_PROPNAME,
-                TransportConstants.TCP_NODELAY_PROPNAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.TCP_SENDBUFFER_SIZE_PROPNAME,
-                TransportConstants.TCP_SENDBUFFER_SIZE_PROPNAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.TCP_RECEIVEBUFFER_SIZE_PROPNAME,
-                TransportConstants.TCP_RECEIVEBUFFER_SIZE_PROPNAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.NIO_REMOTING_THREADS_PROPNAME,
-                TransportConstants.NIO_REMOTING_THREADS_PROPNAME);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.BATCH_DELAY,
-                TransportConstants.BATCH_DELAY);
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.NIO_REMOTING_THREADS_PROPNAME,
-                TransportConstants.NIO_REMOTING_THREADS_PROPNAME);
-        PARAM_KEY_MAPPING.put(
-                ActiveMQDefaultConfiguration.getPropMaskPassword(),
-                HornetQDefaultConfiguration.getPropMaskPassword());
-        PARAM_KEY_MAPPING.put(
-                ActiveMQDefaultConfiguration.getPropPasswordCodec(),
-                HornetQDefaultConfiguration.getPropPasswordCodec());
-        PARAM_KEY_MAPPING.put(
-                org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.NETTY_CONNECT_TIMEOUT,
-                TransportConstants.NETTY_CONNECT_TIMEOUT);
-    }
-
     private final InjectedValue<ActiveMQBroker> injectedActiveMQServer = new InjectedValue<ActiveMQBroker>();
-    private final HornetQConnectionFactory uncompletedConnectionFactory;
+    private final ConnectionFactory uncompletedConnectionFactory;
     private final String discoveryGroupName;
     private final List<String> connectors;
+    private final LegacyConnectionFactory factory;
 
-    private HornetQConnectionFactory connectionFactory;
+    private ConnectionFactory connectionFactory;
 
-    public LegacyConnectionFactoryService(HornetQConnectionFactory uncompletedConnectionFactory, String discoveryGroupName, List<String> connectors) {
-
+    public LegacyConnectionFactoryService(LegacyConnectionFactory factory, ConnectionFactory uncompletedConnectionFactory, String discoveryGroupName, List<String> connectors) {
+        this.factory = factory;
         this.uncompletedConnectionFactory = uncompletedConnectionFactory;
         this.discoveryGroupName = discoveryGroupName;
         this.connectors = connectors;
@@ -156,66 +48,7 @@ public class LegacyConnectionFactoryService implements Service<ConnectionFactory
     @Override
     public void start(StartContext context) throws StartException {
         ActiveMQServer activeMQServer = ActiveMQServer.class.cast(injectedActiveMQServer.getValue().getDelegate());
-
-        DiscoveryGroupConfiguration discoveryGroupConfiguration = null;
-        if (discoveryGroupName != null) {
-            if (activeMQServer.getConfiguration().getDiscoveryGroupConfigurations().containsKey(discoveryGroupName)) {
-                discoveryGroupConfiguration = translateDiscoveryGroupConfiguration(activeMQServer.getConfiguration().getDiscoveryGroupConfigurations().get(discoveryGroupName));
-            } else {
-                throw MessagingLogger.ROOT_LOGGER.discoveryGroupIsNotDefined(discoveryGroupName);
-            }
-        }
-        TransportConfiguration[] transportConfigurations = translateTransportGroupConfigurations(activeMQServer.getConfiguration().getConnectorConfigurations(), connectors);
-        JMSFactoryType factoryType = JMSFactoryType.valueOf(uncompletedConnectionFactory.getFactoryType());
-
-        if (uncompletedConnectionFactory.isHA()) {
-
-            if (discoveryGroupConfiguration != null) {
-                connectionFactory = HornetQJMSClient.createConnectionFactoryWithHA(discoveryGroupConfiguration, factoryType);
-            } else {
-                connectionFactory = HornetQJMSClient.createConnectionFactoryWithHA(factoryType, transportConfigurations);
-            }
-        } else {
-            if (discoveryGroupConfiguration != null) {
-                connectionFactory = HornetQJMSClient.createConnectionFactoryWithoutHA(discoveryGroupConfiguration, factoryType);
-            } else {
-                connectionFactory = HornetQJMSClient.createConnectionFactoryWithoutHA(factoryType, transportConfigurations);
-            }
-        }
-
-        connectionFactory.setAutoGroup(uncompletedConnectionFactory.isAutoGroup());
-        connectionFactory.setBlockOnAcknowledge(uncompletedConnectionFactory.isBlockOnAcknowledge());
-        connectionFactory.setBlockOnDurableSend(uncompletedConnectionFactory.isBlockOnDurableSend());
-        connectionFactory.setBlockOnNonDurableSend(uncompletedConnectionFactory.isBlockOnNonDurableSend());
-        connectionFactory.setCacheLargeMessagesClient(uncompletedConnectionFactory.isCacheLargeMessagesClient());
-        connectionFactory.setCallFailoverTimeout(uncompletedConnectionFactory.getCallFailoverTimeout());
-        connectionFactory.setCallTimeout(uncompletedConnectionFactory.getCallTimeout());
-        connectionFactory.setClientFailureCheckPeriod(uncompletedConnectionFactory.getClientFailureCheckPeriod());
-        connectionFactory.setClientID(uncompletedConnectionFactory.getClientID());
-        connectionFactory.setCompressLargeMessage(uncompletedConnectionFactory.isCompressLargeMessage());
-        connectionFactory.setConfirmationWindowSize(uncompletedConnectionFactory.getConfirmationWindowSize());
-        connectionFactory.setConnectionLoadBalancingPolicyClassName(uncompletedConnectionFactory.getConnectionLoadBalancingPolicyClassName());
-        connectionFactory.setConnectionTTL(uncompletedConnectionFactory.getConnectionTTL());
-        connectionFactory.setConsumerMaxRate(uncompletedConnectionFactory.getConsumerMaxRate());
-        connectionFactory.setConsumerWindowSize(uncompletedConnectionFactory.getConsumerWindowSize());
-        connectionFactory.setConfirmationWindowSize(uncompletedConnectionFactory.getConfirmationWindowSize());
-        connectionFactory.setDupsOKBatchSize(uncompletedConnectionFactory.getDupsOKBatchSize());
-        connectionFactory.setFailoverOnInitialConnection(uncompletedConnectionFactory.isFailoverOnInitialConnection());
-        connectionFactory.setGroupID(uncompletedConnectionFactory.getGroupID());
-        connectionFactory.setInitialConnectAttempts(uncompletedConnectionFactory.getInitialConnectAttempts());
-        connectionFactory.setInitialMessagePacketSize(uncompletedConnectionFactory.getInitialMessagePacketSize());
-        connectionFactory.setMaxRetryInterval(uncompletedConnectionFactory.getMaxRetryInterval());
-        connectionFactory.setMinLargeMessageSize(uncompletedConnectionFactory.getMinLargeMessageSize());
-        connectionFactory.setPreAcknowledge(uncompletedConnectionFactory.isPreAcknowledge());
-        connectionFactory.setProducerMaxRate(uncompletedConnectionFactory.getProducerMaxRate());
-        connectionFactory.setProducerWindowSize(uncompletedConnectionFactory.getProducerWindowSize());
-        connectionFactory.setReconnectAttempts(uncompletedConnectionFactory.getReconnectAttempts());
-        connectionFactory.setRetryInterval(uncompletedConnectionFactory.getRetryInterval());
-        connectionFactory.setRetryIntervalMultiplier(uncompletedConnectionFactory.getRetryIntervalMultiplier());
-        connectionFactory.setScheduledThreadPoolMaxSize(uncompletedConnectionFactory.getScheduledThreadPoolMaxSize());
-        connectionFactory.setThreadPoolMaxSize(uncompletedConnectionFactory.getThreadPoolMaxSize());
-        connectionFactory.setTransactionBatchSize(uncompletedConnectionFactory.getTransactionBatchSize());
-        connectionFactory.setUseGlobalPools(uncompletedConnectionFactory.isUseGlobalPools());
+        connectionFactory = factory.completeConnectionFactory(activeMQServer, uncompletedConnectionFactory, discoveryGroupName, connectors);
     }
 
     @Override
@@ -229,12 +62,13 @@ public class LegacyConnectionFactoryService implements Service<ConnectionFactory
     }
 
     public static LegacyConnectionFactoryService installService(final String name,
-                                                                final ServiceName activeMQServerServiceName,
-                                                                final ServiceTarget serviceTarget,
-                                                                final HornetQConnectionFactory uncompletedConnectionFactory,
-                                                                final String discoveryGroupName,
-                                                                final List<String> connectors) {
-        final LegacyConnectionFactoryService service = new LegacyConnectionFactoryService(uncompletedConnectionFactory, discoveryGroupName, connectors);
+            final ServiceName activeMQServerServiceName,
+            final ServiceTarget serviceTarget,
+            final LegacyConnectionFactory factory,
+            final ConnectionFactory uncompletedConnectionFactory,
+            final String discoveryGroupName,
+            final List<String> connectors) {
+        final LegacyConnectionFactoryService service = new LegacyConnectionFactoryService(factory, uncompletedConnectionFactory, discoveryGroupName, connectors);
         final ServiceName serviceName = JMSServices.getConnectionFactoryBaseServiceName(activeMQServerServiceName).append(LEGACY, name);
 
         final ServiceBuilder sb = serviceTarget.addService(serviceName, service);
@@ -245,71 +79,4 @@ public class LegacyConnectionFactoryService implements Service<ConnectionFactory
         return service;
     }
 
-    private DiscoveryGroupConfiguration translateDiscoveryGroupConfiguration(org.apache.activemq.artemis.api.core.DiscoveryGroupConfiguration newDiscoveryGroupConfiguration) throws StartException {
-        org.apache.activemq.artemis.api.core.BroadcastEndpointFactory newBroadcastEndpointFactory = newDiscoveryGroupConfiguration.getBroadcastEndpointFactory();
-        BroadcastEndpointFactoryConfiguration legacyBroadcastEndpointFactory;
-
-        if (newBroadcastEndpointFactory instanceof org.apache.activemq.artemis.api.core.UDPBroadcastEndpointFactory) {
-            org.apache.activemq.artemis.api.core.UDPBroadcastEndpointFactory factory = (org.apache.activemq.artemis.api.core.UDPBroadcastEndpointFactory) newBroadcastEndpointFactory;
-            legacyBroadcastEndpointFactory = new UDPBroadcastGroupConfiguration(
-                    factory.getGroupAddress(),
-                    factory.getGroupPort(),
-                    factory.getLocalBindAddress(),
-                    factory.getLocalBindPort());
-        } else if (newBroadcastEndpointFactory instanceof org.apache.activemq.artemis.api.core.ChannelBroadcastEndpointFactory) {
-            org.apache.activemq.artemis.api.core.ChannelBroadcastEndpointFactory factory = (org.apache.activemq.artemis.api.core.ChannelBroadcastEndpointFactory) newBroadcastEndpointFactory;
-            legacyBroadcastEndpointFactory = new org.hornetq.api.core.JGroupsBroadcastGroupConfiguration(
-                    factory.getChannel(),
-                    factory.getChannelName());
-        } else {
-            throw MessagingLogger.ROOT_LOGGER.unsupportedBroadcastGroupConfigurationForLegacy(newBroadcastEndpointFactory.getClass().getName());
-        }
-
-        return new DiscoveryGroupConfiguration(newDiscoveryGroupConfiguration.getName(),
-                newDiscoveryGroupConfiguration.getRefreshTimeout(),
-                newDiscoveryGroupConfiguration.getDiscoveryInitialWaitTimeout(),
-                legacyBroadcastEndpointFactory);
-    }
-
-    private TransportConfiguration[] translateTransportGroupConfigurations(Map<String, org.apache.activemq.artemis.api.core.TransportConfiguration> connectorConfigurations, List<String> connectors) throws StartException {
-        List<TransportConfiguration> legacyConnectorConfigurations = new ArrayList<>();
-
-        for (String connectorName : connectors) {
-            org.apache.activemq.artemis.api.core.TransportConfiguration newTransportConfiguration = connectorConfigurations.get(connectorName);
-            String legacyFactoryClassName = translateFactoryClassName(newTransportConfiguration.getFactoryClassName());
-            Map legacyParams = translateParams(newTransportConfiguration.getParams());
-            TransportConfiguration legacyTransportConfiguration = new TransportConfiguration(
-                    legacyFactoryClassName,
-                    legacyParams,
-                    newTransportConfiguration.getName());
-
-            legacyConnectorConfigurations.add(legacyTransportConfiguration);
-        }
-
-        return legacyConnectorConfigurations.toArray(new TransportConfiguration[legacyConnectorConfigurations.size()]);
-    }
-
-    private String translateFactoryClassName(String newFactoryClassName) throws StartException {
-        if (newFactoryClassName.equals(org.apache.activemq.artemis.core.remoting.impl.netty.NettyConnectorFactory.class.getName())) {
-            return org.hornetq.core.remoting.impl.netty.NettyConnectorFactory.class.getName();
-        } else {
-            throw MessagingLogger.ROOT_LOGGER.unsupportedConnectorFactoryForLegacy(newFactoryClassName.getClass().getName());
-        }
-    }
-
-    private Map translateParams(Map<String, Object> newParams) {
-        Map<String, Object> legacyParams = new HashMap<>();
-
-        for (Map.Entry<String, Object> newEntry : newParams.entrySet()) {
-            String newKey = newEntry.getKey();
-            Object value = newEntry.getValue();
-            String legacyKey = PARAM_KEY_MAPPING.getOrDefault(newKey, newKey);
-            if (org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants.ACTIVEMQ_SERVER_NAME.equals(legacyKey)) {
-                // property specific to ActiveMQ that can not be mapped to HornetQ
-                continue;
-            }
-            legacyParams.put(legacyKey, value);
-        }
-        return legacyParams;
-    }
 }
