+++ testsuite/integration/legacy/extra-content/standalone/configuration/application-roles.properties
@@ -1,0 +1,9 @@
+#
+# Copyright The WildFly Authors
+# SPDX-License-Identifier: Apache-2.0
+#
+
+guest=guest
+user1=Users,Role1
+user2=Users,Role2
+# $local=Users,Role1,Role2
