+++ testsuite/integration/legacy/extra-content/standalone/configuration/application-users.properties
@@ -1,0 +1,13 @@
+#
+# Copyright The WildFly Authors
+# SPDX-License-Identifier: Apache-2.0
+#
+
+# guest=guest
+# user1=password1
+# user2=password2
+#remoteejbuser=rem@teejbpasswd1
+guest=b5d048a237bfd2874b6928e1f37ee15e
+user1=23624d2f74dfcb9688651a066d90b97e
+user2=ab3f9e12039435236d89de9023a304b7
+remoteejbuser=d37cd830cc282510807b82c4b861256d
